small fixes

+++ modules/binary/src/main/c++/cli/optionParser.c++
@@ -57,7 +57,7 @@ namespace TyrantGenetic {
                     throw InvalidUserInputError("Need a deck to mutate.");
                 }
 
-                configuration.verbosity = vm.count("verbose");
+                configuration.verbosity = static_cast<int>(vm.count("verbose"));
 
                 RunCommand::Ptr command = RunCommand::Ptr(
                     new RunCommand(configuration)
